Update: refactor structure
