+++ src/main/common/utils/taskList/tasks/confirmYourDetails.ts
@@ -1,6 +1,6 @@
-import {Task} from "../../../common/form/models/task";
-import {Claim} from "../../../common/models/claim";
-import {TaskStatus} from "../../../common/form/models/TaskStatus";
+import {Task} from "../../../models/taskList/task";
+import {Claim} from "../../../models/claim";
+import {TaskStatus} from "../../../models/taskList/TaskStatus";
 
 /**
  * THIS FILE IS A CONCEPT AND DOESN'T WORK
