+++ src/main/views/features/response/task-list.njk
@@ -18,7 +18,7 @@
             {{ t('After you have completed all the actions you will be taken to a page where you can check your answers before submitting.') }}
           </p>
 
-          {% for taskList in data %}
+          {% for taskList in taskLists %}
             <div class="{{ 'govuk-!-display-none' if taskList.tasks.length===0}}">
               <h2 class="app-task-list__section">
                 <span class="app-task-list__section-number">
@@ -41,7 +41,7 @@
           {% endfor %}
 
           <h2 class="app-task-list__section">
-            {{ data.length + 1 }}. Submit
+            {{ taskLists.length + 1 }}. Submit
           </h2>
           <ol class="app-task-list app-task-list__items">
             <form method="post">
